[PHP] Add coverage reporting
Adds test coverage reporting to the composer `phpunit` script as a
way of surfacing basic coverage information. The coverage report is
printed to stdout to keep things lightweight.

Also added is a `coverage-html` script which writes an HTML coverage
report in the docs/ subdirectory.

Hygiene: Update MetricsClientTest to apply the @covers tag at the
class scope as recommended in the docs
(https://phpunit.readthedocs.io/en/9.5/code-coverage-analysis.html#specifying-covered-code-parts).

Bug: T288742
Change-Id: Iee6fcaf086209440f0a05c65b8535a7a5da381d7