[Java] Restore support for multiple destination event services
Restores full support for sending events to different intake services
based on the stream config. Concretely, the library now supports
submitting error events to eventgate-logging-external.

Note: This commit drops the inputBuffer/outputBuffer terminology in
favor of more explanatory names: unvalidatedEvents, validatedEvents,
and validatedErrors. The latter are implemented as instances of a
new EventBuffer class which associates an ArrayList of events with
a destination event service.

Bug: T290770
Change-Id: I3bf14dfac78f676f95c5bbc230eaa700aaa24ca9