[Swift] Add support for multiple destination event services
Adds support for sending events to different intake services
based on the stream config. With this change, the library supports
submitting error events to eventgate-logging-external.

Note: This commit drops the inputBuffer/outputBuffer terminology in
favor of more explanatory names: unvalidatedEvents, validatedEvents,
and validatedErrors. The latter are implemented as instances of a
new EventBuffer class which associates a queue of events with a
destination event service.

Bug: T290770
Change-Id: I71466b11f6c3b9f7fbd8c6b15684d92ec1459b49