[JS] Add MetricsClientIntegration::onSubmit()
If the user has enabled the the debug mode of the EventLogging
extension, then they will see a notification or JavaScript console entry
for every event submitted with mw.eventLog.submit().

In order the Metrics Platform client to support that behaviour, we call
MetricsClientIntegration::onSubmit() after an event has been enqueued
for submission. Since the behaviour is platform-specific, make
::onSubmit() do nothing by default.

Bug: T281761
Change-Id: I9fddcdaa26b221723880c5fc264abf14445d701a