[PHP]: Validate stream configuration
We can simplify MetricsClient, ContextController, and CurationController
by:

1. Creating an API to interact with the stream configuration; and

2. Extracting all validation of that stream configuration to a separate
collaborator.

For (1), we introduce StreamConfig, a class that encapsulates how to
fetch those parts of the stream configuration that direct
ContextController and CurationController.

For (2), we introduce StreamConfigFactory and
ValidatingStreamConfigFactory. The former establishes an API to get the
configuration for the given stream and does very simple validation of
that configuration, whereas the latter validates the stream
configuration against a JSON Schema.

The motivation for having both light- and heavyweight validation is to
allow codebases to perform heavyweight validation at integration time
and lightweight validation at runtime.

Bug: T281762
Change-Id: Ida0abdd0f1dd241194423a15ab5d8d76bc035140