[JS] Reduce library size
In [0], Krinkle highlighted that the Metrics Platform JavaScript
implementation increases the size of the ext.eventLogging ResourceLoader
module by 200+%.

The most obvious ways to reduce the size of the library are:

1. Removing the repetitious switch statements in
   ContextController::addRequestedValues() and
   CurationController::shouldProduceEvent(); and

2. Removing the IMetricsClientIntegration implementation and replace it
   with a TypeScript type declaration

We introduce the Integration TypeScript interface, which has the
smallest possible set of methods on the old IMetricsClientInterface
implementation. We then add Integration::getContextAttributes(), which
returns all possible context attributes that are available in the
execution environment. Those context attributes are in the same shape as
they are on on an event so that they can be accessed in the same way.

[0] https://gerrit.wikimedia.org/r/c/mediawiki/extensions/EventLogging/+/754002/24#message-43465166d964bd77f8d44a4c9f85e7590531faa5

Bug: T281761
Change-Id: I2a1bf0860fce67d2b0fbba2c3f06b2b6058bd78d