Include /js/ in the Git archive
In I3953ec4bc35, we use ForeignResourceManager from MediaWiki Core to
download the Git archive and extract all JavaScript files. However,
I1fe7493d574 excluded those files from the Git archive.

In order to support both use cases for the Git archive, include /js/ in
the Git archive. This means that Composer will download the directory
as well but this is an acceptable cost as it enables an
easier-to-maintain solution in at least the EventLogging extension.

Change-Id: Ia30205fda8c15e09285b560a7bf68b3ae6e09764