Make EventSender typesafe.
Instead of passing a String that represents a URL to
`EventSender.sendEvents()`, we can pass a URL directly.

This implies a few changes to make the rest of the code typesafe as
well, but those are minor and I think improve readability by
making more explicit that we are passing around URLs and not
generic strings.

Change-Id: If92d9e719919a49239a96a03d5d3bfcfba09e152