Cleanup EventProcessorTest
EventProcessorTest now is less reliant on CurationFilter, which
makes this test less flaky. It uses a mock instead of a real
CurationFilter and only relies on the CurationFilter passing
or not passing, not on its internal behaviour.

As part of this cleanup, some fixtures have been extracted to
their own class, so that they can be reused between tests.

Change-Id: Ief98531298c950e8bf193ce6b4705e51d84ef3f8