[Java] Allow page metadata to be submitted with event
The 1st release of the MP Java client library makes assumptions
about client-originated metadata that would be available at the
instantiation of the MetricsClient object, namely page metadata.

This proposed 2nd release posits that page metadata is dynamic
(as it is in the case of the android app, the main client for
this library) similar to custom data and should be added at
the time of event submission (as opposed to injected as an
implementing class in the MetricsClient constructor).

This patch also separates concerns between the event that is
submitted by the client and the eventual event that is handled
by the EventProcessor. The latter includes processing the
original Event object with additional metadata that is
immutable and injected into the MetricsClient at the time of
creation.

The following changes are introduced accordingly:

- Create EventProcessed child class from Event.
- Update Event to only contain custom + page data.
- Overload submitMetricsEvent() with PageData parameters.
- Replace ClientMetadata interface with ClientData object.
- Add method to create PageData from map.
- Add helper methods to create PageData from json.
- Update relevant tests and ClientMetadata implementations.

Bug: T330355
Change-Id: Ib99779dabe389d5de7c81f1ee4d2e1230ef9258b