[Java] Remove unnecessary usage of Lombok.NonNull annotation.
This removes a single (accidental?) usage of Lombok.NonNull,
replacing it with javax.Nonnull. This annotation was leaking into
the final compiled library, and was requiring us to explicitly
ignore it in our build optimizer.

This also updates a couple of methods to no longer return boxed
booleans unnecessarily, which trips up my bug checker.

And a general question: it looks like you're using the
`addLombokGeneratedAnnotation` configuration parameter, which adds
some other annotations that we need to explicitly ignore. What's
the rationale for that parameter, and can it be removed too?

Change-Id: I3182ee5d439521191b9692a8d4f75b3a995279ce