[Java] Always send enqueued events through thread executor.
Currently the onAppPause() and onAppClose() methods send the
enqueued events immediately, on the same thread as the function
is called. Since those functions will be called from the main
thread on Android, this will cause the app to crash.

This modifies the logic to always call the sendEnqueuedEvents
function through our thread executor service.

Change-Id: Ia2df128e5c05a31a207cc4d08072164943cdf47f