[Java] Correctly serialize groups of events when sending.
When sending more than one event at a time, the library was
incorrectly serializing them one after another, which creates an invalid
JSON structure, and results in a 500 response from the server.

This updates the serialization to act directly on the collection of
events, which also has the benefit of being even simpler.

Change-Id: I786d4622ed42e271e38eb51d318fff097b122940