Ensure that client data nested data objects are built fresh
For clients that submit optional, overriding client data with an
event, we need to make sure the client data object of the event is
not mixed up during batch processing of the event queue. During the
filtering of contextual values in the ContextController, we rebuild
the client data object for each event when stream config specifies
requested values in its producer config.

This patch updates the ContextController to return fresh client
data when the event queue is processed.

This patch also adds a new value for the local stream config to match
the expected event ("agent_app_install_id").

Bug: T338964
Change-Id: Ia41473a495d8e56bbdc1ffd0e0d809618c837d3c