[Java] Improve and refine error handling.
* The response from the server will NOT be 200 (OK); it will be 201
for the non-hasty calls, and 202 for hasty calls. The library was
therefore incorrectly showing warnings for successful calls.
* When the server returns error 4xx, it will likely contain a helpful
message body that details how exactly the event(s) are malformed.
Let's capture this message in an exception and log it loudly.
* Remove code for logging success message from the server, which
there won't be anyway.
* Change the name of the Hasty switch from isProd to isDebug. It's
not only "production" releases that will use the hasty parameter;
the Beta app will use it, too. It's slightly more accurate to say that
the non-hasty url is used for *debugging* purposes.

Change-Id: Icce98d44c554d3e1901e02225de7302d5cac14b3