[Swift] Use a tagged Docker image instead of latest
The Docker image for Swift does not mention any tag, it thus assumes
`latest` but CI does not pull the image to get a fresher version.

Also if a new image is build, that can introduces unexpected changes, so
it is probably better to use an explicity tag.

Follow up rebuild of docker-registry.discovery.wmnet/dev/buster-swift53:0.2.0
https://gitlab.wikimedia.org/repos/releng/dev-images/-/merge_requests/51/

Change-Id: I65bcf52870e2d8f5c65190e8cae6da8f063bf99d