Move SampleConfig to its own package to prevent cycles.
The next commit will introduce package cycles if SampleConfig isn't
moved to its own package. It might make sense to dig a bit deeper
into the packages semantics here and define more clearly what is
configuration and what is code to load and manage the configuration,
but this might come at another step.