Allow MetricsClient#newInstrument(instrumentName)
* [JS] Bump web base schema ID to v1.3.0

* [JS] Document MetricsClient#newInstrument()

* [JS] MetricsClient#newInstrument(): Use default stream

[0] introduced the 'product_metrics.web_base' stream. This stream is
intended to be the default stream for all events submitted by the
Metrics Platform Clients unless the Instrument Developer is using their
own schema. This allows us to simplify MetricsClient#newInstrument().

Previously, if the stream name couldn't be determined from the
instrument name, then we would log a warning and return an instance of
NullInstrument. However, we can now default to using the
'product_metrics.web_base' stream.

Changes:

* Make MetricsClient#newInstrument() default to using the
  'product_metrics.web_base' stream
* Remove NullInstrument

[0]: https://gerrit.wikimedia.org/r/c/operations/mediawiki-config/+/1074396

* [JS] Allow MetricsClient#newInstrument( instrumentName )

* [JS] Add TODO about StreamProducerConfig.sampling