[JS] Allow an event to be redirected to another stream
What?

If the config for stream S1 has the
producers.metrics_platform_client.stream_name property that refers to
another stream S2 and S2 is configured, then check that S1 is in-sample
and, if so, submit the event to S2.

Why?

xLab UI allows analytics instrument owners to specify a target stream.
The target stream may product_metrics.web_base or it may be something
else.

In all other ways, instruments and streams are configured the same way
and therefore can be modelled using the StreamConfig structure. However,
submit() needs to be tweaked slightly to handle the difference in
behavior transparently.

Bug: T401380